rename box3 translateBy to offset
