+++ lib/box3/index.js
@@ -25,6 +25,7 @@ exports.getSphere = require('./getSphere')
 exports.getVolume = require('./getVolume')
 exports.hasPoint = require('./hasPoint')
 exports.homothety = require('./homothety')
+exports.offset = require('./offset')
 exports.projectToPlane = require('./projectToPlane')
 exports.projectTo = exports.projectToPlane
 exports.resizeBy = require('./resizeBy')
@@ -33,5 +34,5 @@ exports.rotateBy = require('./rotateBy')
 exports.scaleBy = exports.homothety
 exports.transitFrom = require('./transitFrom')
 exports.transitTo = require('./transitTo')
-exports.translateBy = require('./translateBy')
+exports.translateBy = exports.offset
 exports.validate = require('./validate')
