+++ test/box3/index.test.js
@@ -16,13 +16,13 @@ const units = {
   getVolume: require('./getVolume.test'),
   hasPoint: require('./hasPoint.test'),
   homothety: require('./homothety.test'),
+  offset: require('./offset.test'),
   projectToPlane: require('./projectToPlane.test'),
   resizeBy: require('./resizeBy.test'),
   resizeTo: require('./resizeTo.test'),
   rotateBy: require('./rotateBy.test'),
   transitFrom: require('./transitFrom.test'),
   transitTo: require('./transitTo.test'),
-  translateBy: require('./translateBy.test'),
   validate: require('./validate.test')
 }
